Started subdividing movegenerator

+++ AI.cpp
@@ -6,7 +6,7 @@ AI::AI(string FEN, color computerColor)
 	genChessData data;
 	data.gen(); // Generates bitboards needed for move generation
 	chessBoard = Board(FEN);
-	transposition_hash = ZobristHash(1e7);
+	transposition_hash = ZobristHash(static_cast<size_t>(1e7));
 	//debug();
 }
